refactor : 초기 postService 분리작업
