+++ src/main/java/inu/codin/codin/domain/post/service/PostQueryService.java
@@ -1,0 +1,148 @@
+package inu.codin.codin.domain.post.service;
+
+import inu.codin.codin.common.exception.NotFoundException;
+import inu.codin.codin.common.security.util.SecurityUtils;
+import inu.codin.codin.domain.block.service.BlockService;
+import inu.codin.codin.domain.like.entity.LikeType;
+import inu.codin.codin.domain.post.domain.poll.service.PollService;
+import inu.codin.codin.domain.post.dto.UserDto;
+import inu.codin.codin.domain.post.dto.UserInfo;
+import inu.codin.codin.domain.post.dto.response.PollInfoResponseDTO;
+import inu.codin.codin.domain.post.dto.response.PostDetailResponseDTO;
+import inu.codin.codin.domain.post.dto.response.PostPageItemResponseDTO;
+import inu.codin.codin.domain.post.dto.response.PostPageResponse;
+import inu.codin.codin.domain.post.entity.PostCategory;
+import inu.codin.codin.domain.post.entity.PostEntity;
+import inu.codin.codin.domain.post.repository.PostRepository;
+import inu.codin.codin.domain.user.entity.UserEntity;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.bson.types.ObjectId;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class PostQueryService
+{
+    private final PostRepository postRepository;
+    private final PollService pollService;
+    private final BlockService blockService;
+    private final SeperatedPostService seperatedPostService;
+
+    /**
+     * 카테고리별 삭제되지 않은 게시물 목록 조회
+     * @return PostPageResponse (불변 리스트)
+     */
+    public PostPageResponse getAllPosts(PostCategory postCategory, int pageNumber) {
+        List<ObjectId> blockedUsersId = blockService.getBlockedUsers();
+        PageRequest pageRequest = PageRequest.of(pageNumber, 20, Sort.by("createdAt").descending());
+        Page<PostEntity> page = postRepository.getPostsByCategoryWithBlockedUsers(postCategory.toString(), blockedUsersId, pageRequest);
+        log.info("모든 글 반환 성공 Category: {}, Page: {}", postCategory, pageNumber);
+        return PostPageResponse.of(getPostListResponseDtos(page.getContent()), page.getTotalPages() - 1, page.hasNext() ? page.getPageable().getPageNumber() + 1 : -1);
+    }
+
+    /**
+     * 게시물 리스트 DTO 변환 (불변 리스트)
+     */
+    public List<PostPageItemResponseDTO> getPostListResponseDtos(List<PostEntity> posts) {
+        return posts.stream()
+                .map(this::toPageItemDTO)
+                .toList();
+    }
+
+    /**
+     * 게시물 상세 조회
+     */
+    public PostPageItemResponseDTO getPostWithDetail(String postId) {
+        PostEntity post = postRepository.findByIdAndNotDeleted(new ObjectId(postId))
+                .orElseThrow(() -> new NotFoundException("게시물을 찾을 수 없습니다. id=" + postId));
+        ObjectId userId = SecurityUtils.getCurrentUserId();
+        seperatedPostService.increaseHitsIfNeeded(post, userId);
+        return toPageItemDTO(post);
+    }
+
+    /**
+     * Optional로 게시물 상세 조회 (null-safe)
+     */
+    public Optional<PostPageItemResponseDTO> getPostDetailById(ObjectId postId) {
+        return postRepository.findByIdAndNotDeleted(postId)
+                .map(post -> {
+                    ObjectId userId = SecurityUtils.getCurrentUserId();
+                    seperatedPostService.increaseHitsIfNeeded(post, userId);
+                    return toPageItemDTO(post);
+                });
+    }
+
+    /**
+     * 키워드 기반 게시물 검색
+     */
+    public PostPageResponse searchPosts(String keyword, int pageNumber) {
+        List<ObjectId> blockedUsersId = blockService.getBlockedUsers();
+        PageRequest pageRequest = PageRequest.of(pageNumber, 20, Sort.by("createdAt").descending());
+        Page<PostEntity> page = postRepository.findAllByKeywordAndDeletedAtIsNull(keyword, blockedUsersId, pageRequest);
+        log.info("키워드 기반 게시물 검색: {}, Page: {}", keyword, pageNumber);
+        return PostPageResponse.of(getPostListResponseDtos(page.getContent()), page.getTotalPages() - 1, page.hasNext() ? page.getPageable().getPageNumber() + 1 : -1);
+    }
+
+    /**
+     * Top 3 베스트 게시물 조회 (불변 리스트)
+     */
+    public List<PostPageItemResponseDTO> getTop3BestPosts() {
+        List<PostEntity> bestPosts = seperatedPostService.getTop3BestPostsInternal();
+        log.info("Top 3 베스트 게시물 반환.");
+        return getPostListResponseDtos(bestPosts);
+    }
+
+    /**
+     * 베스트 게시물 페이지 조회
+     */
+    public PostPageResponse getBestPosts(int pageNumber) {
+        Page<PostEntity> page = seperatedPostService.getBestPostsInternal(pageNumber);
+        return PostPageResponse.of(getPostListResponseDtos(page.getContent()), page.getTotalPages() - 1, page.hasNext() ? page.getPageable().getPageNumber() + 1 : -1);
+    }
+
+
+
+    // PostEntity → PostPageItemResponseDTO 변환 (공통 변환 로직)
+    private PostPageItemResponseDTO toPageItemDTO(PostEntity post) {
+        UserDto userDto = resolveUserProfile(post);
+        int likeCount = seperatedPostService.getLikeCount(post);
+        int scrapCount = seperatedPostService.getScrapCount(post);
+        int hitsCount = seperatedPostService.getHitsCount(post);
+        int commentCount = post.getCommentCount();
+        ObjectId userId = SecurityUtils.getCurrentUserId();
+        UserInfo userInfo = getUserInfoAboutPost(userId, post.getUserId(), post.get_id());
+        PostDetailResponseDTO postDTO = PostDetailResponseDTO.of(post, userDto, likeCount, scrapCount, hitsCount, commentCount, userInfo);
+        if (post.getPostCategory() == PostCategory.POLL) {
+            PollInfoResponseDTO pollInfo = pollService.getPollInfo(post, userId);
+            return PostPageItemResponseDTO.of(postDTO, pollInfo);
+        } else {
+            return PostPageItemResponseDTO.of(postDTO, null);
+        }
+    }
+
+    // [유저 프로필] - 닉네임/이미지 결정
+    private UserDto resolveUserProfile(PostEntity post) {
+        UserEntity user = userRepository.findById(post.getUserId())
+                .orElseThrow(() -> new NotFoundException("유저를 찾을 수 없습니다."));
+        return UserDto.ofPost(post, user, s3Service.getDefaultProfileImageUrl());
+    }
+
+    // [유저 프로필] - 게시물에 대한 유저정보 추출
+    public UserInfo getUserInfoAboutPost(ObjectId currentUserId, ObjectId postUserId, ObjectId postId){
+        return UserInfo.of(
+                likeService.isLiked(LikeType.POST, postId, currentUserId),
+                scrapService.isPostScraped(postId, currentUserId),
+                postUserId.equals(currentUserId)
+        );
+    }
+
+
+}
